Adapted student model for postgres, added integrity checks to its DAO

+++ src/model/student.py
@@ -1,7 +1,7 @@
 # Arquivo de mapeamento objeto-relacional para estudantes.
 
 from sqlalchemy.orm import relationship
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, String, Integer, BigInteger
 from model import Base
 
 
@@ -11,7 +11,7 @@ class Student(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String)
     registry = Column(Integer)          # matrícula
-    discord_id = Column(Integer, nullable=False, unique=True)
+    discord_id = Column(BigInteger, nullable=False, unique=True)
 
     is_assigned = relationship('Assigned', back_populates='student')
     registered_on = relationship('Registered', back_populates='student')
